requirements.txt
# Deep Learning Framework - PyTorch
torch>=2.0.0
torchvision>=0.15.0  # Optional, for additional utilities

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data Loading
ijson>=3.2.0  # For streaming large JSON files

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Metrics and Evaluation
scipy>=1.10.0

# Optional but recommended for GPU
# Install CUDA-compatible PyTorch from: https://pytorch.org/get-started/locally/
# Example: pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu118
